requirements.txt
terratorch>=0.99
rasterio>=1.3
rioxarray>=0.15
odc-stac>=0.3
pystac-client>=0.7
planetary-computer>=1.0
pyproj>=3.6
numpy>=1.26
pandas>=2.1
pyyaml>=6.0
tqdm>=4.66
scikit-learn>=1.4
